+++ test/jest/__mock__/stripesSmartComponent.mock.js
@@ -1,4 +1,9 @@
 jest.mock('@folio/stripes/smart-components', () => ({
   ...jest.requireActual('@folio/stripes/smart-components'),
-  ConfigManager: jest.fn(() => <div>ConfigManager</div>),
+  ConfigManager: jest.fn(({ lastMenu }) => (
+    <div>
+      ConfigManager
+      {lastMenu}
+    </div>
+  )),
 }));