+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 11.0.0 IN PROGRESS
 
 * *BREAKING* Add `Language & Localization` setting. Refs UIMPROF-107.
+* User preference: Select a language > Reset to tenant default language. Refs UIMPROF-115.
 
 ## 10.0.0 (https://github.com/folio-org/ui-myprofile/tree/v10.0.0) (2025-03-14)
 [Full Changelog](https://github.com/folio-org/ui-myprofile/compare/v9.2.0...v10.0.0)