+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change history for ui-myprofile
 
+## IN PROGRESS
+
 ## 11.0.0 (https://github.com/folio-org/ui-myprofile/tree/v11.0.0) (2026-04-16)
 [Full Changelog](https://github.com/folio-org/ui-myprofile/compare/v10.0.0...v11.0.0)